requirements.txt
# Core dependencies
tika==2.6.0
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
spacy==3.7.2
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.3.0
click==8.1.7

# Support libraries
pandas==2.0.3
pydantic==2.3.0
numpy==1.24.4
requests==2.31.0

# OCR language packs - installation note: requires 'python -m spacy download en_core_web_lg'

# Dev dependencies
pytest==7.4.4
black==24.1.1
pylint==2.17.6
flake8==6.1.0
pytest-cov==4.1.0 